refactor: extract connection and query logic to custom hook
